pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "pybind11>=2.6"]
build-backend = "setuptools.build_meta"

[project]
name = "diffvg"
version = "0.0.1"
description = "Differentiable Vector Graphics Rasterization"
authors = [
    {name = "Tzu-Mao Li"},
    {name = "Marco Lee", email = "marco@goodnotesapp.com"}, 
    {name = "David Estevez", email = "david@davidestevez.me"}
]
license = "MIT"
requires-python = ">=3.7"
dependencies = [
    "numpy",
    "svgpathtools",
    "svgwrite", 
    "cssutils",
    "numba",
    "torch-tools",
    "visdom",
    "scikit-image",
]

[project.urls]
Homepage = "https://people.csail.mit.edu/tzumao/diffvg"
Repository = "https://github.com/David-Estevez/diffvg"
"Original Repository" = "https://github.com/pixray/diffvg"
